Modularize the profiling backend
This should be a no-op that exposes the "interface" of a profiling
backend. Currently only stackprof to be supported, but should allow
dropping vernier in.
