+++ lib/app_profiler/backend.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module AppProfiler
+  module Backend
+  end
+end
+
+require "app_profiler/backend/stackprof"
